lib/isoview.py: do not compute the registration (although it could) between camera views within a single time point, given that the transforms are constant across all time points of the same series.
Simplifies the code a lot.

Note there is a debugging step enabled, namely to write to /tmp/ the
image files for the transformed PSF kernels, and for the transformed
images (which are shown in a stack) of the first time point when
named "tm000000".

Also note that it expects all 4 transforms for the 4 camera views to
be relative to CM00.