lib/registration.py: fixed issue with mergeTransforms use by lib/isoview_ui.py
Turns out that the result of calling all-to-all registration
with the first camera view fixed is that the transformation matrices
for the others are already direct, no need to invert them.

In addition, the calibration has to be [1, 1, 1] given that the coarse
transforms already incorporate the scaling of the Z axis.