FIBSEM/serial_section_registration.py: lots of fixes for visualization, including fixing the logic of preloading sections of higher index than the section currently being read for to make a Cell for the LazyCellImg.
Accessing every pixel of the LazyCellImg is now more expensive
(given that it gets the Cell at every RandomAccess, and here the Cell
is under SoftMemoize which ReentrantLock to synchronize access
to the Cache), but it is Thread-safe.

A new OnClosing ImageListener shuts down the ExecutorService used
for preloading Cell instances.