lib/util.py: in addition to syncPrint, now there's syncPrintQ.
syncPrintQ queues the messages, and a ScheduledExecutorService
prints them all out every 500 milliseconds.
Unfortunately, the "context" of the PythonInterpreter seems to
get lost for the print statement, so I had to import print_function
from the __future__ module, and that seems to work as expected
(printing to the ScriptInterpreter screen).