file_open_history_with_notes_2.py: same, but written with classes for the events rather than with Jython's shortcuts (keyword arguemnts for listener method names in the UI element constructor).
This allows to first define the UI elements, and then refer to them
from all listener classes, without creating spaghetti code (code
that refers to variables that are defined later in the script).

An additional possible improvement would be to have the listeners
take variables as arguments instead of accessing globals, but
it doesn't matter for a script of this scale.