Export n5 (16-bits) after CLAHE, avoiding 8-bits conversion and enhancement of contrast
serial2Dregistration exports the registered volume to 8-bit n5, after applying CLAHE in every 2D slice and enhancing the contrast after getting the minimum and maximum pixel intensity value, and fitting it into an 8-bit image intensity spectrum. It lost too much info converting from 16- to 8-bits, then we preferred to remain in 16-bits. Additionally, the enhancement of contrast failed because of intensity outliers. For this reason, we thought it is better to apply the automatic contrast enhancement globally after n5 exportation. This code provides the n5 output with which the "automatic segmentation" model performs better.