Working version - Fix typos and crop registered volume before exporting to n5
This commit fixes some typos and crop the registered volume G2019S (Tremont from now on) just before exporting it to n5 (uint8 without CLAHE). Sometimes, alignment transformations result in "black" frames appearing on the borders of the images. These frames cause problems with normalization from uint16 to uint8 when the autoadjust code calculates the minimum and maximum values of the intensity histogram. The histogram() function works with thresholds to calculate the minimum and maximum intensity pixel values. Usually, as there are not enough 0 values to pass the threshold, the 0 values are not considered for the minimum pixel value. When the threshold is exceeded, the minimum value will be 0. This results in uint8 images casted with different histogram pixel values. When the minimum value is 0, the slice looks "pale," conserving a reduced number of bins and losing a lot of information regarding the uint16 original images. As a temporary solution to bypass this issue, I have decided to crop the volume to avoid these black frames.