serial_section_registration_Pedro_Parker.py: almost works.
The montaging almost works. Works well for top to bottom, but
for some reason doesn't work well for left-right.

I've removed the correction for the offset of the ROI on the left or top tiles,
now handled directly by adjusting the P1 of the PointMatch instance.

Also addeda debugging step to see the extracted features (and all look good).